rename to fluent-json-schema
